better macro replace
smart {.tag} which stops the sub macro evaluation at that level, so tag can contain valid {tag} codes that will never be replaces

smart {-tag} which evaluates to nothing if the chunk or the first level of references within the chunk do not exist.

so {.test} where test="{test}" would evaluate to "{test}" without any recursion problems or further substitutions

so {-test} where test="this is {it}" will evaluate to "" if either test or it are not valid chunks.

This gives a simple if style logic test for the most common substitution case.